pyproject.toml
[project]
name = "mcp-tutorial"
version = "0.1.0"
description = "MCP tutorial project"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.51.0",
    "arxiv>=2.2.0",
    "mcp>=1.9.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
]
